+++ src/features/create-post/components/next-step-link/index.ts
@@ -1,0 +1,1 @@
+export * from './next-step-link'
