+++ src/features/create-post/steps/filters/filters.tsx
@@ -6,17 +6,15 @@ import { useWizard } from 'react-use-wizard'
 
 import style from './filters.module.scss'
 
-import ImageFilter from '@/features/create-post/components/image-filter/image-filter'
-import { NextStepLink } from '@/features/create-post/components/next-step-link/next-step-link'
+import { ImageFilter, NextStepLink } from '@/features/create-post/components'
 import { filterNames } from '@/features/create-post/constants/canvas-filters'
-import { useImageCropContext } from '@/features/create-post/context/crop-provider'
 import { CloseModal } from '@/features/create-post/steps/close-modal/close-modal'
-import { DotsBar } from '@/features/create-post/ui/dots-bar/dots-bar'
-import NewPostModal from '@/features/create-post/ui/new-post-modal/new-post-modal'
+import { DotsBar, NewPostModal } from '@/features/create-post/ui'
 import { useSlider } from '@/features/create-post/utils/use-slider'
 import { ArrowBack2 } from '@/shared/assets/icons/arrow-back-icon/arrow-back2'
 import next from '@/shared/assets/icons/filter-post-photo/next.svg'
 import prev from '@/shared/assets/icons/filter-post-photo/prev.svg'
+import { useImageCropContext } from '@/shared/hooks/use-image-crop-context'
 import { Button, ButtonTheme } from '@/shared/ui'
 
 export const Filters = () => {
