+++ src/features/create-post/steps/publication/publication.tsx
@@ -8,16 +8,16 @@ import { useWizard } from 'react-use-wizard'
 import style from './publication.module.scss'
 
 import { NextStepLink } from '@/features/create-post/components/next-step-link/next-step-link'
-import { useImageCropContext } from '@/features/create-post/context/crop-provider'
 import { CloseModal } from '@/features/create-post/steps/close-modal/close-modal'
 import { ImagePublication } from '@/features/create-post/steps/image-publication/image-publication'
-import NewPostModal from '@/features/create-post/ui/new-post-modal/new-post-modal'
+import { NewPostModal } from '@/features/create-post/ui/new-post-modal/new-post-modal'
 import {
   useCreatePostMutation,
   useUploadImageMutation,
 } from '@/shared/api/services/posts/posts.api'
 import { useGetProfileUserQuery } from '@/shared/api/services/profile/profile.api'
 import { ArrowBack2 } from '@/shared/assets/icons/arrow-back-icon/arrow-back2'
+import { useImageCropContext } from '@/shared/hooks/use-image-crop-context'
 import { LinearLoader, Input, InputType } from '@/shared/ui'
 
 export const Publication = () => {
