+++ src/features/create-post/ui/dots-bar/index.ts
@@ -1,0 +1,1 @@
+export * from './dots-bar'
