+++ src/features/create-post/ui/error-image-message/error-message-image.module.scss
