+++ src/features/create-post/ui/error-image-message/index.ts
@@ -1,0 +1,1 @@
+export * from './error-message-image'
