+++ src/features/create-post/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './dots-bar'
+export * from './new-post-modal'
+export * from './error-image-message'
