+++ src/features/create-post/ui/new-post-modal/index.ts
@@ -1,0 +1,1 @@
+export * from './new-post-modal'
