+++ src/features/post/ui/close-edit-modal/index.ts
@@ -1,0 +1,1 @@
+export * from './close-edit-modal'
