+++ src/features/post/ui/edit-delete-post/index.ts
@@ -1,0 +1,1 @@
+export * from './edit-delete-post'
