+++ src/features/post/ui/edit-description-modal/index.ts
@@ -1,0 +1,1 @@
+export * from './edit-description-post'
