+++ src/features/post/ui/index.ts
@@ -1,0 +1,4 @@
+export * from './edit-delete-post/'
+export * from './close-edit-modal/'
+export * from './edit-description-modal/'
+export * from './post-images-modal/'
