+++ src/features/post/ui/post-images-modal/index.ts
@@ -1,0 +1,1 @@
+export * from './post-images'
