+++ src/features/super-admin-user-management/ui/user-ban-modal/user-ban-modal.tsx
@@ -6,7 +6,7 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import s from './user-ban-modal.module.scss'
 
-import NewPostModal from '@/features/create-post/ui/new-post-modal/new-post-modal'
+import { NewPostModal } from '@/features/create-post/ui/'
 import { useBanUserMutation } from '@/features/super-admin-user-management/lib/handle-user-ban'
 import {
   selectBanModalOpenStatus,
