+++ src/features/super-admin-user-management/ui/user-delete-modal/user-delete-modal.tsx
@@ -6,7 +6,7 @@ import { useDispatch, useSelector } from 'react-redux'
 
 import s from './user-delete-modal.module.scss'
 
-import NewPostModal from '@/features/create-post/ui/new-post-modal/new-post-modal'
+import { NewPostModal } from '@/features/create-post/ui/new-post-modal/new-post-modal'
 import { useDeleteUserMutation } from '@/features/super-admin-user-management/lib/handle-user-delete'
 import {
   selectDeleteModalOpenStatus,
